requirements.txt
Flask>=3.1.0
marker-pdf>=1.7.5
markdown>=3.7